refactor instructions mod, some clippy cleanup
