+++ boring-vault-svm-core/src/instructions/mod.rs
@@ -1,0 +1,9 @@
+pub mod boring_vault;
+pub mod jito;
+pub mod solend;
+pub mod system;
+
+pub use boring_vault::*;
+pub use jito::*;
+pub use solend::*;
+pub use system::*;
