+++ boring-vault-svm-core/src/manage_instructions/jito.rs
@@ -1,6 +1,6 @@
+use crate::manage_instructions::ExternalInstruction;
 use crate::utils::bindings::boring_vault_svm::types::{Operator, Operators};
 use crate::utils::pdas;
-use crate::manage_instructions::ExternalInstruction;
 
 use solana_instruction::account_meta::AccountMeta;
 use solana_program::system_program;
@@ -10,7 +10,8 @@ use spl_token::ID as TOKEN_PROGRAM_ID;
 
 const JITO_MINT: Pubkey = pubkey!("J1toso1uCk3RLmjorhTtrVwY9HJ7X8V9yYac6Y7kGCPn");
 const JITO_STAKE_POOL: Pubkey = pubkey!("Jito4APyf642JPZPx3hGc6WWJ8zPKtRbRs4P815Awbb");
-const JITO_STAKE_POOL_WITHDRAW_AUTHORITY: Pubkey = pubkey!("6iQKfEyhr3bZMotVkW6beNZz5CPAkiwvgV2CTje9pVSS");
+const JITO_STAKE_POOL_WITHDRAW_AUTHORITY: Pubkey =
+    pubkey!("6iQKfEyhr3bZMotVkW6beNZz5CPAkiwvgV2CTje9pVSS");
 const JITO_RESERVE_STAKE_ACCOUNT: Pubkey = pubkey!("BgKUXdS29YcHCFrPm5M8oLHiTzZaMDjsebggjoaQ6KFL");
 const JITO_FEE_ACCOUNT: Pubkey = pubkey!("feeeFLLsam6xZJFc6UQFrHqkvVt4jfmVvi2BRLkUZ4i");
 
@@ -56,28 +57,16 @@ impl ExternalInstruction for MintJitoSol {
             get_associated_token_address_with_program_id(&vault_pda, &JITO_MINT, &TOKEN_PROGRAM_ID);
 
         vec![
-            AccountMeta::new(
-                JITO_STAKE_POOL,
-                false,
-            ), // Jito (JitoSol) Stake Pool
-            AccountMeta::new_readonly(
-                JITO_STAKE_POOL_WITHDRAW_AUTHORITY,
-                false,
-            ), // stake pool withdraw authority
-            AccountMeta::new(
-                JITO_RESERVE_STAKE_ACCOUNT,
-                false,
-            ), // reserve stake account
-            AccountMeta::new(vault_pda, false),    // depositor
-            AccountMeta::new(jito_sol_ata, false), // user account
-            AccountMeta::new(
-                JITO_FEE_ACCOUNT,
-                false,
-            ), // fee account
-            AccountMeta::new(jito_sol_ata, false), // referral fee account
-            AccountMeta::new(JITO_MINT, false),         // token mint
+            AccountMeta::new(JITO_STAKE_POOL, false), // Jito (JitoSol) Stake Pool
+            AccountMeta::new_readonly(JITO_STAKE_POOL_WITHDRAW_AUTHORITY, false), // stake pool withdraw authority
+            AccountMeta::new(JITO_RESERVE_STAKE_ACCOUNT, false), // reserve stake account
+            AccountMeta::new(vault_pda, false),                  // depositor
+            AccountMeta::new(jito_sol_ata, false),               // user account
+            AccountMeta::new(JITO_FEE_ACCOUNT, false),           // fee account
+            AccountMeta::new(jito_sol_ata, false),               // referral fee account
+            AccountMeta::new(JITO_MINT, false),                  // token mint
             AccountMeta::new_readonly(system_program::ID, false), // system program
-            AccountMeta::new_readonly(TOKEN_PROGRAM_ID, false), // token program
+            AccountMeta::new_readonly(TOKEN_PROGRAM_ID, false),  // token program
         ]
     }
