+++ boring-vault-svm-core/src/manage_instructions/system.rs
@@ -1,6 +1,6 @@
+use crate::manage_instructions::ExternalInstruction;
 use crate::utils::bindings::boring_vault_svm::types::{Operator, Operators};
 use crate::utils::pdas;
-use crate::manage_instructions::ExternalInstruction;
 use solana_instruction::account_meta::AccountMeta;
 use solana_program::system_instruction;
 use solana_program::system_program;
